+++ packages/integrations/node/src/preview-entrypoint.ts
@@ -1,0 +1,45 @@
+import { fileURLToPath } from 'node:url';
+import { AstroError } from 'astro/errors';
+import logListeningOn from './log-listening-on.js';
+import { createServer } from './standalone.js';
+import type { CreatePreviewServer } from 'astro';
+import type { createExports } from './server-entrypoint.js';
+import type { RequestHandler } from './types.js';
+
+type ServerModule = ReturnType<typeof createExports>;
+type MaybeServerModule = Partial<ServerModule>;
+
+const preview: CreatePreviewServer = async function (preview) {
+	let ssrHandler: ServerModule['handler'];
+	let options: ServerModule['options'];
+	try {
+		process.env.ASTRO_NODE_AUTOSTART = 'disabled';
+		const ssrModule: MaybeServerModule = await import(preview.serverEntrypoint.toString());
+		if (typeof ssrModule.handler === 'function') {
+			ssrHandler = ssrModule.handler;
+			options = ssrModule.options!;
+		} else {
+			throw new AstroError(
+				`The server entrypoint doesn't have a handler. Are you sure this is the right file?`
+			);
+		}
+	} catch (err) {
+		if ((err as any).code === 'ERR_MODULE_NOT_FOUND') {
+			throw new AstroError(
+				`The server entrypoint ${fileURLToPath(
+					preview.serverEntrypoint
+				)} does not exist. Have you ran a build yet?`
+			);
+		} else {
+			throw err;
+		}
+	}
+	const host = preview.host ?? "127.0.0.1"
+	const port = preview.port ?? 4321
+	const server = createServer(ssrHandler as RequestHandler, host, port);
+	logListeningOn(preview.logger, server.server, options)
+	server.server.listen(port, host);
+	return server;
+};
+
+export { preview as default };
