+++ packages/integrations/node/src/server-entrypoint.ts
@@ -1,7 +1,8 @@
-import type { SSRManifest } from 'astro';
 import { NodeApp, applyPolyfills } from 'astro/app/node';
-import middleware from './nodeMiddleware.js';
+import { createStandaloneHandler } from './standalone.js';
 import startServer from './standalone.js';
+import createMiddleware from './middleware.js';
+import type { SSRManifest } from 'astro';
 import type { Options } from './types.js';
 
 applyPolyfills();
@@ -9,7 +10,10 @@ export function createExports(manifest: SSRManifest, options: Options) {
 	const app = new NodeApp(manifest);
 	return {
 		options: options,
-		handler: middleware(app, options.mode),
+		handler:
+			options.mode === "middleware"
+				? createMiddleware(app)
+				: createStandaloneHandler(app, options),
 		startServer: () => startServer(app, options),
 	};
 }
